+++ precompiles/staking/events.go
@@ -22,6 +22,8 @@ import (
 const (
 	// EventTypeCreateValidator defines the event type for the staking CreateValidator transaction.
 	EventTypeCreateValidator = "CreateValidator"
+	// EventTypeEditValidator defines the event type for the staking DeitValidator transaction.
+	EventTypeEditValidator = "EditValidator"
 	// EventTypeDelegate defines the event type for the staking Delegate transaction.
 	EventTypeDelegate = "Delegate"
 	// EventTypeUnbond defines the event type for the staking Undelegate transaction.
@@ -142,6 +144,41 @@ func (p Precompile) EmitCreateValidatorEvent(ctx sdk.Context, stateDB vm.StateDB
 	// Prepare the event data
 	var b bytes.Buffer
 	b.Write(cmn.PackNum(reflect.ValueOf(msg.Value.Amount.BigInt())))
+
+	stateDB.AddLog(&ethtypes.Log{
+		Address:     p.Address(),
+		Topics:      topics,
+		Data:        b.Bytes(),
+		BlockNumber: uint64(ctx.BlockHeight()),
+	})
+
+	return nil
+}
+
+// EmitEditValidatorEvent creates a new edit valdator event emitted on a EditValidator transaction.
+func (p Precompile) EmitEditValidatorEvent(ctx sdk.Context, stateDB vm.StateDB, msg *stakingtypes.MsgEditValidator, validatorAddr common.Address) error {
+	// Prepare the event topics
+	event := p.ABI.Events[EventTypeEditValidator]
+
+	topics, err := p.createEditValidatorTxTopics(2, event, validatorAddr)
+	if err != nil {
+		return err
+	}
+
+	commissionRate := big.NewInt(DoNotModifyCommissionRate)
+	if msg.CommissionRate != nil {
+		commissionRate = msg.CommissionRate.BigInt()
+	}
+
+	minSelfDelegation := big.NewInt(DoNotModifyMinSelfDelegation)
+	if msg.MinSelfDelegation != nil {
+		minSelfDelegation = msg.MinSelfDelegation.BigInt()
+	}
+
+	// Prepare the event data
+	var b bytes.Buffer
+	b.Write(cmn.PackNum(reflect.ValueOf(commissionRate)))
+	b.Write(cmn.PackNum(reflect.ValueOf(minSelfDelegation)))
 
 	stateDB.AddLog(&ethtypes.Log{
 		Address:     p.Address(),
@@ -308,6 +345,24 @@ func (p Precompile) createCreateValidatorTxTopics(topicsLen uint64, event abi.Ev
 	}
 
 	topics[2], err = cmn.MakeTopic(validatorAddr)
+	if err != nil {
+		return nil, err
+	}
+
+	return topics, nil
+}
+
+// createEditValidatorTxTopics creates the topics for staking transactions EditValidator.
+func (p Precompile) createEditValidatorTxTopics(topicsLen uint64, event abi.Event, validatorAddr common.Address) ([]common.Hash, error) {
+	topics := make([]common.Hash, topicsLen)
+	// NOTE: If your solidity event contains indexed event types, then they become a topic rather than part of the data property of the log.
+	// In solidity you may only have up to 4 topics but only 3 indexed event types. The first topic is always the signature of the event.
+
+	// The first topic is always the signature of the event.
+	topics[0] = event.ID
+
+	var err error
+	topics[1], err = cmn.MakeTopic(validatorAddr)
 	if err != nil {
 		return nil, err
 	}