+++ precompiles/staking/tx.go
@@ -5,6 +5,7 @@ package staking
 
 import (
 	"fmt"
+	"strconv"
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -20,6 +21,8 @@ import (
 const (
 	// CreateValidatorMethod defines the ABI method name for the staking create validator transaction
 	CreateValidatorMethod = "createValidator"
+	// EditValidatorMethod defines the ABI method name for the staking edit validator transaction
+	EditValidatorMethod = "editValidator"
 	// DelegateMethod defines the ABI method name for the staking Delegate
 	// transaction.
 	DelegateMethod = "delegate"
@@ -81,8 +84,59 @@ func (p Precompile) CreateValidator(
 		return nil, err
 	}
 
-	// Emit the event for the delegate transaction
+	// Emit the event for the create validator transaction
 	if err = p.EmitCreateValidatorEvent(ctx, stateDB, msg, delegatorHexAddr); err != nil {
+		return nil, err
+	}
+
+	return method.Outputs.Pack(true)
+}
+
+// EditValidator performs edit validator.
+func (p Precompile) EditValidator(
+	ctx sdk.Context,
+	origin common.Address,
+	_ *vm.Contract,
+	stateDB vm.StateDB,
+	method *abi.Method,
+	args []interface{},
+) ([]byte, error) {
+	msg, validatorHexAddr, err := NewMsgEditValidator(args)
+	if err != nil {
+		return nil, err
+	}
+
+	commissionRate := strconv.Itoa(DoNotModifyCommissionRate)
+	if msg.CommissionRate != nil {
+		commissionRate = msg.CommissionRate.String()
+	}
+
+	minSelfDelegation := strconv.Itoa(DoNotModifyMinSelfDelegation)
+	if msg.CommissionRate != nil {
+		commissionRate = msg.MinSelfDelegation.String()
+	}
+
+	p.Logger(ctx).Debug(
+		"tx called",
+		"method", method.Name,
+		"validator_address", msg.ValidatorAddress,
+		"commission_rate", commissionRate,
+		"min_self_delegation", minSelfDelegation,
+	)
+
+	// we only allow the tx signer "origin" to edit their own validator.
+	if origin != validatorHexAddr {
+		return nil, fmt.Errorf(ErrDifferentOriginFromValidator, origin.String(), validatorHexAddr.String())
+	}
+
+	// Execute the transaction using the message server
+	msgSrv := stakingkeeper.NewMsgServerImpl(&p.stakingKeeper)
+	if _, err = msgSrv.EditValidator(sdk.WrapSDKContext(ctx), msg); err != nil {
+		return nil, err
+	}
+
+	// Emit the event for the edit validator transaction
+	if err = p.EmitEditValidatorEvent(ctx, stateDB, msg, validatorHexAddr); err != nil {
 		return nil, err
 	}